requirements.txt
numpy
pytorch
scikit-learn
matplotlib
